create favorite function in book detail

+++ app/Http/Controllers/CommentController.php
@@ -6,6 +6,10 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use App\Models\Comment;
 
+use function Symfony\Component\String\u;
+
+use App\Models\User;
+
 class CommentController extends Controller
 {
     private $comment;
@@ -15,6 +19,7 @@ class CommentController extends Controller
        
     }
 
+    // コメント投稿
     public function store(Request $request, $post_id) {
          $request->validate([
              'comment_body' . $post_id => 'required|max:150'
@@ -30,13 +35,24 @@ class CommentController extends Controller
              'post_id'  => $post_id
          ]);
     
-         return redirect()->route('user.posts.show', $post_id);
+         return redirect()->route('user.posts.show', $post_id)
+             ->with('comment_posted',true);
     } 
 
+    // コメント削除
     public function destroy($id) {
+        $comment = $this->comment->findOrFail($id);
+        if(Auth::user()->id != $comment->user_id) {
+            abort(403);
+        }
        $this->comment->destroy($id);
-       return redirect()->back();
+       return redirect()->back()->with('comment_deleted',true);
     }
-       
+
+    // ユーザーアバター取得
+    public function getUserAvatar($user_id) {
+        $user = User::find($user_id);
+        return $user->avatar ? asset('storage/avatars/' . $user->avatar) : asset('images/default-avatar.png');
+    }
     
 }
